+++ projects/angular-intl-demo/src/app/date/date.component.ts
@@ -1,4 +1,5 @@
-import { Component } from '@angular/core';
+import {Component} from '@angular/core';
+import {languages} from '../languages';
 
 @Component({
   selector: 'app-date',
@@ -6,8 +7,10 @@ import { Component } from '@angular/core';
   styleUrls: ['./date.component.scss']
 })
 export class DateComponent {
+  languages = languages;
   selectedDate = new Date();
   dateStyle: Intl.DateTimeFormatOptions['dateStyle'];
   timeStyle: Intl.DateTimeFormatOptions['timeStyle'];
   hour12: Intl.DateTimeFormatOptions['hour12'];
+  locale?: string;
 }