+++ projects/angular-intl-demo/src/app/language/language.component.ts
@@ -1,5 +1,5 @@
 import {Component} from '@angular/core';
-import {languages} from "./languages";
+import {languages} from "../languages";
 import {IntlLanguagePipeOptions} from "projects/angular-ecmascript-intl/src/lib/language/intl-language.pipe";
 
 @Component({
@@ -11,4 +11,5 @@ export class LanguageComponent {
   languages = languages;
   selectedLanguage = 'de-DE';
   languageDisplay?: IntlLanguagePipeOptions['languageDisplay'];
+  locale?: string;
 }